+++ cannon/mipsevm/program/metadata.go
@@ -8,8 +8,8 @@ import (
 
 type Symbol struct {
 	Name  string `json:"name"`
-	Start uint32 `json:"start"`
-	Size  uint32 `json:"size"`
+	Start uint64 `json:"start"`
+	Size  uint64 `json:"size"`
 }
 
 type Metadata struct {
@@ -27,12 +27,12 @@ func MakeMetadata(elfProgram *elf.File) (*Metadata, error) {
 	})
 	out := &Metadata{Symbols: make([]Symbol, len(syms))}
 	for i, s := range syms {
-		out.Symbols[i] = Symbol{Name: s.Name, Start: uint32(s.Value), Size: uint32(s.Size)}
+		out.Symbols[i] = Symbol{Name: s.Name, Start: uint64(s.Value), Size: uint64(s.Size)}
 	}
 	return out, nil
 }
 
-func (m *Metadata) LookupSymbol(addr uint32) string {
+func (m *Metadata) LookupSymbol(addr uint64) string {
 	if len(m.Symbols) == 0 {
 		return "!unknown"
 	}
@@ -50,19 +50,19 @@ func (m *Metadata) LookupSymbol(addr uint32) string {
 	return out.Name
 }
 
-type SymbolMatcher func(addr uint32) bool
+type SymbolMatcher func(addr uint64) bool
 
 func (m *Metadata) CreateSymbolMatcher(name string) SymbolMatcher {
 	for _, s := range m.Symbols {
 		if s.Name == name {
 			start := s.Start
 			end := s.Start + s.Size
-			return func(addr uint32) bool {
+			return func(addr uint64) bool {
 				return addr >= start && addr < end
 			}
 		}
 	}
-	return func(addr uint32) bool {
+	return func(addr uint64) bool {
 		return false
 	}
 }
